fetch_sqs_msg/requirements.txt

boto3==1.38.38
botocore==1.38.38
greenlet==3.2.3
jmespath==1.0.1
playwright==1.52.0
pyee==13.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
s3transfer==0.13.0
six==1.17.0
typing-extensions==4.14.0
urllib3==2.5.0
